Handle null when getting parameter from object property.
If you define a prameter that gets its value from the property of object the template will fall over if you don't provide the object when calling the template.
this will allow you to not need to pass the object every time.
example --@recruitedByPersonId Int = recruitedBy.personId
Currently if you don't post recruitedBy the template will fail to run.
 in this case we check if rerutiedByPersonId is null in the sql so we dont' need to pass the object .